Cleanup "main.rs"
